1241: Mark juleica applications different
